master: added branch for routes fix

+++ routes/admin/web.php
@@ -63,7 +63,7 @@ Route::group(['prefix' => 'admin', 'middleware' => ['auth', 'admin'], 'namespace
         Route::post('/', 'QuestionsController@getQuestions');
         Route::post('/create', 'QuestionsController@create');
         Route::post('/load-image', 'QuestionsController@loadImage');
-        Route::delete('/{question}', 'QuestionsController@delete');
+        Route::delete('/{question}', 'QuestionsController@delete')->name('admin.questions.delete');
     });
 
     Route::group(['prefix' => 'lessons-settings', 'namespace' => 'Settings'], function () {
